+++ crates/client/sync/src/lib.rs
@@ -73,6 +73,7 @@ pub mod starknet_sync_worker {
             telemetry,
         ),)?;
 
+
         Ok(())
     }
 }